+++ src/opt.rs
@@ -1,6 +1,6 @@
 use std::fmt::Debug;
 use crate::ast;
-use crate::ast::{Program,Error,Expression,FunctionDefinition,LIPart,LHSPart,LHSLiteralPart};
+use crate::ast::{Program,Error,Expression};
 use cranelift::prelude::*;
 use cranelift_jit::{JITBuilder, JITModule};
 use cranelift_module::{DataContext, Linkage, Module};
@@ -112,181 +112,7 @@ impl JProgram {
 pub fn check_hardcoded_call<'f, S: Clone + Debug>(ctx: &mut FunctionBuilder<'f>, p: &Program<S>, fi: usize, args: &Vec<(JExpr,JType)>) -> Option<(JExpr,JType)> {
    let sig = args.iter().map(|(_je,jt)| jt.jtype).collect::<Vec<types::Type>>();
    let val = args.iter().map(|(je,_jt)| je.value).collect::<Vec<Value>>();
-   let hardcoded: Vec<(Vec<types::Type>,FunctionDefinition<()>,fn(&mut FunctionBuilder<'f>,Vec<Value>) -> Value)> = vec![
-      (vec![types::I64,types::I64],
-       FunctionDefinition::define(
-         vec![0,1],
-         vec![Expression::li(vec![
-            LIPart::variable(0),
-            LIPart::variable(1),
-         ],())]
-      ),|ctx,val| {
-         let val0 = val[0].clone();
-         let val1 = val[1].clone();
-         ctx.ins().iadd(val0, val1)
-      }),
-      (vec![types::I64,types::I64],
-       FunctionDefinition::define(
-          vec![0,1],
-          vec![Expression::pattern(
-             Expression::variable(0,()),
-             vec![(
-                LHSPart::ul(
-                   vec![LHSLiteralPart::variable(1)],
-                   Some(2),
-                   vec![],
-                ),
-                Expression::variable(2,()),
-             )],
-          ())],
-      ),|ctx,val| {
-         let val0 = val[0].clone();
-         let val1 = val[1].clone();
-         ctx.ins().isub(val0, val1)
-      }),
-      (vec![types::I64,types::I64],
-       FunctionDefinition::define(
-          vec![0,1],
-          vec![Expression::pattern(
-             Expression::variable(0,()),
-             vec![(
-                LHSPart::ul(
-                   vec![LHSLiteralPart::variable(1)],
-                   None,
-                   vec![],
-                ),
-                Expression::unary(b"1",()),
-             ),(
-                LHSPart::Any,
-                Expression::unary(b"0",()),
-             )],
-          ())],
-      ),|ctx,val| {
-         let val0 = val[0].clone();
-         let val1 = val[1].clone();
-         let vi8 = ctx.ins().icmp(IntCC::Equal, val0, val1);
-         ctx.ins().uextend(types::I64, vi8)
-      }),
-      (vec![types::I64,types::I64],
-       FunctionDefinition::define(
-          vec![0,1],
-          vec![Expression::pattern(
-             Expression::variable(0,()),
-             vec![(
-                LHSPart::ul(
-                   vec![LHSLiteralPart::variable(1)],
-                   None,
-                   vec![],
-                ),
-                Expression::unary(b"0",()),
-             ),(
-                LHSPart::Any,
-                Expression::unary(b"1",()),
-             )],
-          ())],
-      ),|ctx,val| {
-         let val0 = val[0].clone();
-         let val1 = val[1].clone();
-         let vi8 = ctx.ins().icmp(IntCC::NotEqual, val0, val1);
-         ctx.ins().uextend(types::I64, vi8)
-      }),
-      (vec![types::I64,types::I64],
-       FunctionDefinition::define(
-          vec![0,1],
-          vec![Expression::pattern(
-             Expression::variable(1,()),
-             vec![(
-                LHSPart::ul(
-                   vec![LHSLiteralPart::variable(0),
-                        LHSLiteralPart::literal("0")],
-                   Some(2),
-                   vec![],
-                ),
-                Expression::unary(b"1",()),
-             ),(
-                LHSPart::Any,
-                Expression::unary(b"0",()),
-             )],
-         ())],
-      ),|ctx,val| {
-         let val0 = val[0].clone();
-         let val1 = val[1].clone();
-         let vi8 = ctx.ins().icmp(IntCC::UnsignedLessThan, val0, val1);
-         ctx.ins().uextend(types::I64, vi8)
-      }),
-      (vec![types::I64,types::I64],
-       FunctionDefinition::define(
-          vec![0,1],
-          vec![Expression::pattern(
-             Expression::variable(0,()),
-             vec![(
-                LHSPart::ul(
-                   vec![LHSLiteralPart::variable(1),
-                        LHSLiteralPart::literal("0")],
-                   Some(2),
-                   vec![],
-                ),
-                Expression::unary(b"1",()),
-             ),(
-                LHSPart::Any,
-                Expression::unary(b"0",()),
-             )],
-         ())],
-      ),|ctx,val| {
-         let val0 = val[0].clone();
-         let val1 = val[1].clone();
-         let vi8 = ctx.ins().icmp(IntCC::UnsignedGreaterThan, val0, val1);
-         ctx.ins().uextend(types::I64, vi8)
-      }),
-      (vec![types::I64,types::I64],
-       FunctionDefinition::define(
-          vec![0,1],
-          vec![Expression::pattern(
-             Expression::variable(1,()),
-             vec![(
-                LHSPart::ul(
-                   vec![LHSLiteralPart::variable(0),
-                        LHSLiteralPart::literal("0")],
-                   Some(2),
-                   vec![],
-                ),
-                Expression::unary(b"0",()),
-             ),(
-                LHSPart::Any,
-                Expression::unary(b"1",()),
-             )],
-         ())],
-      ),|ctx,val| {
-         let val0 = val[0].clone();
-         let val1 = val[1].clone();
-         let vi8 = ctx.ins().icmp(IntCC::UnsignedGreaterThanOrEqual, val0, val1);
-         ctx.ins().uextend(types::I64, vi8)
-      }),
-      (vec![types::I64,types::I64],
-       FunctionDefinition::define(
-          vec![0,1],
-          vec![Expression::pattern(
-             Expression::variable(0,()),
-             vec![(
-                LHSPart::ul(
-                   vec![LHSLiteralPart::variable(1),
-                        LHSLiteralPart::literal("0")],
-                   Some(2),
-                   vec![],
-                ),
-                Expression::unary(b"0",()),
-             ),(
-                LHSPart::Any,
-                Expression::unary(b"1",()),
-             )],
-         ())],
-      ),|ctx,val| {
-         let val0 = val[0].clone();
-         let val1 = val[1].clone();
-         let vi8 = ctx.ins().icmp(IntCC::UnsignedLessThanOrEqual, val0, val1);
-         ctx.ins().uextend(types::I64, vi8)
-      }),
-   ];
+   let hardcoded = crate::recipes::cranelift::import();
    for (hsig,hdef,hexpr) in hardcoded.iter() {
       if &sig == hsig && p.functions[fi].equals(hdef) {
          let rval = hexpr(ctx, val);
