+++ src/recipes/mod.rs
@@ -1,0 +1,5 @@
+#[cfg(feature = "cranelift")]
+pub mod cranelift;
+
+#[cfg(feature = "cranelift")]
+pub mod cranelift_impl;
